# High performance logging