--- a/src/NexusMods.App.UI/HighPerformanceLogging.cs
+++ b/src/NexusMods.App.UI/HighPerformanceLogging.cs
@@ -1,0 +1,14 @@
+using Microsoft.Extensions.Logging;
+
+namespace NexusMods.App.UI;
+
+internal static partial class HighPerformanceLogging
+{
+    [LoggerMessage(
+        EventId = 0,
+        Level = LogLevel.Debug,
+        Message = "Finding View for {viewModel}")]
+    public static partial void FindingView(
+        this ILogger logger,
+        string viewModel);
+}